Update README.md
The ATM Simulator is a C program that works like a real ATM machine. It allows a user to enter a PIN to log in and then perform basic banking actions like checking balance, depositing money, withdrawing cash, and transferring money to another account. The program also keeps a record of all transactions made by the user. It shows messages if the user tries to withdraw more than the available balance or enters the wrong PIN too many times. This project is useful for learning how to use C language features like if statements, loops, functions, and structures.